refactor: moved anm plugin into anm module
