+++ io_scene_xray/anm/exp.py
@@ -1,9 +1,9 @@
 import io
 import bpy
-from .xray_io import ChunkedWriter, PackedWriter
-from .fmt_anm import Chunks
-from .xray_envelope import export_envelope, EPSILON
-from .utils import smooth_euler
+from ..xray_io import ChunkedWriter, PackedWriter
+from .fmt import Chunks
+from ..xray_envelope import export_envelope, EPSILON
+from ..utils import smooth_euler
 
 
 def _export(bpy_obj, chunked_writer):
