+++ io_scene_xray/xray_inject.py
@@ -1,4 +1,3 @@
-
 import bpy
 
 from .details.types import (
